Done: 1. Fixing a linux compile-time warning.
Need to:
1. Check whether we should add more macro #if checks for tuning, when loading the TM, for speed.
3. Make the number of TM features a static field to optimize memory consumption when tuning.

+++ inc/server/server_configs.hpp
@@ -75,9 +75,9 @@ namespace uva {
 
                 namespace rm {
                     //Stores the number of the reordering model features
-                    static constexpr size_t MAX_NUM_RM_FEATURES = EIGHT_RM_FEATURES;
+                    static constexpr int8_t MAX_NUM_RM_FEATURES = EIGHT_RM_FEATURES;
                     //Stores the half of the maximum number of the reordering model features
-                    static constexpr size_t HALF_MAX_NUM_RM_FEATURES = MAX_NUM_RM_FEATURES / 2;
+                    static constexpr int8_t HALF_MAX_NUM_RM_FEATURES = MAX_NUM_RM_FEATURES / 2;
                 }
             }
         }
